fix(norm): sep var_careate_var_info(), from declare_arg.c to var_create_var_info.c
